+++ src/config/index.ts
@@ -6,7 +6,8 @@ import YAML from "yaml";
 import { setupFunction } from "../setup";
 import { readFileAsync, writeFileAsync } from "../utils/fs/async";
 import * as logger from "../utils/logger";
-import { configPath, mergeMissingProperties, removeUnknownProperties } from "../utils/misc";
+import { mergeMissingProperties, removeUnknownProperties } from "../utils/misc";
+import { configPath } from "../utils/path";
 import defaultConfig from "./default";
 import { IConfig, isValidConfig } from "./schema";
 import { validateConfigExtra } from "./validate";
