+++ src/index.ts
@@ -13,7 +13,7 @@ import startServer from "./server";
 import Image from "./types/image";
 import * as logger from "./utils/logger";
 import { printMaxMemory } from "./utils/mem";
-import { libraryPath } from "./utils/misc";
+import { libraryPath } from "./utils/path";
 
 function skipImage(image: Image) {
   if (!image.path) {
