+++ src/plugins/events/movie.ts
@@ -11,7 +11,8 @@ import Movie from "../../types/movie";
 import Studio from "../../types/studio";
 import { downloadFile } from "../../utils/download";
 import * as logger from "../../utils/logger";
-import { libraryPath, validRating } from "../../utils/misc";
+import { validRating } from "../../utils/misc";
+import { libraryPath } from "../../utils/path";
 import { extensionFromUrl } from "../../utils/string";
 
 // This function has side effects
@@ -23,7 +24,7 @@ export async function onMovieCreate(movie: Movie, event = "movieCreated"): Promi
     movieName: movie.name,
     $createLocalImage: async (path: string, name: string, thumbnail?: boolean) => {
       path = resolve(path);
-      logger.log("Creating image from " + path);
+      logger.log(`Creating image from ${path}`);
       if (await Image.getImageByPath(path)) {
         logger.warn(`Image ${path} already exists in library`);
         return null;
@@ -33,7 +34,7 @@ export async function onMovieCreate(movie: Movie, event = "movieCreated"): Promi
         img.name += " (thumbnail)";
       }
       img.path = path;
-      logger.log("Created image " + img._id);
+      logger.log(`Created image ${img._id}`);
       await imageCollection.upsert(img._id, img);
       if (!thumbnail) {
         await indexImages([img]);
@@ -42,7 +43,7 @@ export async function onMovieCreate(movie: Movie, event = "movieCreated"): Promi
     },
     $createImage: async (url: string, name: string, thumbnail?: boolean) => {
       // if (!isValidUrl(url)) throw new Error(`Invalid URL: ` + url);
-      logger.log("Creating image from " + url);
+      logger.log(`Creating image from ${url}`);
       const img = new Image(name);
       if (thumbnail) {
         img.name += " (thumbnail)";
@@ -51,7 +52,7 @@ export async function onMovieCreate(movie: Movie, event = "movieCreated"): Promi
       const path = libraryPath(`images/${img._id}${ext}`);
       await downloadFile(url, path);
       img.path = path;
-      logger.log("Created image " + img._id);
+      logger.log(`Created image ${img._id}`);
       await imageCollection.upsert(img._id, img);
       if (!thumbnail) {
         await indexImages([img]);
@@ -129,7 +130,7 @@ export async function onMovieCreate(movie: Movie, event = "movieCreated"): Promi
       movie.studio = studio._id;
       await studioCollection.upsert(studio._id, studio);
       await indexStudios([studio]);
-      logger.log("Created studio " + studio.name);
+      logger.log(`Created studio ${studio.name}`);
     }
   }
