+++ src/queue/check.ts
@@ -9,7 +9,7 @@ import Image from "../types/image";
 import Scene from "../types/scene";
 import { statAsync, walk } from "../utils/fs/async";
 import * as logger from "../utils/logger";
-import { libraryPath } from "../utils/misc";
+import { libraryPath } from "../utils/path";
 import ora = require("ora");
 
 export async function checkVideoFolders(): Promise<void> {
@@ -54,7 +54,7 @@ export async function checkVideoFolders(): Promise<void> {
     } catch (error) {
       const _err = error as Error;
       logger.log(_err.stack);
-      logger.error("Error when importing " + videoPath);
+      logger.error(`Error when importing ${videoPath}`);
       logger.warn(_err.message);
     }
   }
@@ -189,7 +189,7 @@ export async function checkPreviews(): Promise<void> {
         const preview = await Scene.generatePreview(scene);
 
         if (preview) {
-          const image = new Image(scene.name + " (preview)");
+          const image = new Image(`${scene.name} (preview)`);
           const stats = await statAsync(preview);
           image.path = preview;
           image.scene = scene._id;
@@ -202,7 +202,7 @@ export async function checkPreviews(): Promise<void> {
           scene.thumbnail = image._id;
           await sceneCollection.upsert(scene._id, scene);
 
-          loader.succeed("Generated preview for " + scene._id);
+          loader.succeed(`Generated preview for ${scene._id}`);
         } else {
           loader.fail(`Error generating preview.`);
         }
