+++ src/types/scene.ts
@@ -22,7 +22,8 @@ import { indexScenes } from "../search/scene";
 import { mkdirpSync, readdirAsync, rimrafAsync, statAsync, unlinkAsync } from "../utils/fs/async";
 import { generateHash } from "../utils/hash";
 import * as logger from "../utils/logger";
-import { generateTimestampsAtIntervals, libraryPath } from "../utils/misc";
+import { generateTimestampsAtIntervals } from "../utils/misc";
+import { libraryPath } from "../utils/path";
 import { removeExtension } from "../utils/string";
 import Actor from "./actor";
 import ActorReference from "./actor_reference";
