+++ src/utils/misc.ts
@@ -1,7 +1,3 @@
-import { readFileSync } from "fs";
-import * as path from "path";
-
-import { getConfig } from "../config";
 import * as logger from "./logger";
 import { isNumber } from "./types";
 
@@ -34,26 +30,6 @@ export function isValidUrl(str: string): boolean {
     logger.error(err);
     return false;
   }
-}
-
-export function libraryPath(str: string): string {
-  return path.join(getConfig().persistence.libraryPath, "library", str);
-}
-
-export const isDocker = (function (): boolean {
-  try {
-    return readFileSync("/proc/self/cgroup", "utf8").includes("docker");
-  } catch (_) {
-    // Will only have error if not in a docker environment
-    return false;
-  }
-})();
-
-export function configPath(...paths: string[]): string {
-  if (isDocker) {
-    return path.resolve(path.join("/config", ...paths));
-  }
-  return path.resolve(process.cwd(), ...paths);
 }
 
 /**
